backend/requirements.txt
flask
flask-cors
gunicorn
psycopg2-binary
requests
ollama
dotenv
cloudinary
flask_sqlalchemy
openai
playwright
flask[async]


